+++ _test_common/lib/runner_asset_writer_spy.dart
@@ -20,4 +20,7 @@ class RunnerAssetWriterSpy extends AssetWriterSpy implements RunnerAssetWriter {
     _assetsDeleted.add(id);
     return _delegate.delete(id);
   }
+
+  @override
+  Future<void> completeBuild() async {}
 }